move get_prompt_with_cache to prompts
